requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==1.30.5
anthropic==0.34.0
cohere==5.9.0
pymupdf==1.24.9
python-docx==1.1.2
pgvector==0.2.5
sqlalchemy==2.0.32
streamlit==1.38.0
requests==2.32.3
httpx==0.27.0
apscheduler==3.10.4
